+++ src/components/radio-button/radio-button.js
@@ -24,12 +24,23 @@ import { tagComponent } from '../../utils/helpers/tags';
  *
  * For additional properties specific to this component, see propTypes.
  *
+ * If you wish to control the checked value of the radio buttons externally to the component
+ * you can use the prop of `checked` e.g.
+ *
+ * <RadioButton name='frequency' value='weekly' checked={ myCheckValue === 'weekly' } label='Weekly' />
+ *
  * @class RadioButton
  * @constructor
  * @decorators {Input, InputLabel, InputValidation}
  */
 const RadioButton = Input(InputLabel(InputValidation(
 class RadioButton extends React.Component {
+
+  static propTypes = {
+    reverse: React.PropTypes.bool,
+    checked: React.PropTypes.bool
+  }
+
   /**
    * Uses the mainClasses method provided by the decorator to add additional classes.
    *
@@ -62,6 +73,7 @@ class RadioButton extends React.Component {
   get fieldHelpClasses() {
     return classNames(
       'carbon-radio-button__help-text', {
+        'carbon-radio-button__help-text--reverse': this.props.reverse,
         'carbon-radio-button__help-text--inline': this.props.fieldHelpInline
       }
     );
@@ -127,11 +139,27 @@ class RadioButton extends React.Component {
    * @return {Object} JSX
    */
   render() {
+    let labelLeft, fieldHelpLeft,
+        labelRight = this.labelHTML,
+        fieldHelpRight = this.fieldHelpHTML;
+
+    if (this.props.reverse) {
+      labelLeft = this.labelHTML;
+      labelRight = null;
+
+      if (this.props.fieldHelpInline) {
+        fieldHelpLeft = this.fieldHelpHTML;
+        fieldHelpRight = null;
+      }
+    }
+
     return(
       <div className={ this.mainClasses } { ...tagComponent('radio-button', this.props) }>
+        { labelLeft }
+        { fieldHelpLeft }
         { this.inputHTML }
-        { this.labelHTML }
-        { this.fieldHelpHTML }
+        { labelRight }
+        { fieldHelpRight }
         { this.validationHTML }
       </div>
     );